Fix Visual Studio warnings at level /W4 (#2312)
* SharePtr fix for gcc 7.2

* Fix for Poco issue #2285

* Add missing fail() to sqlite test

* Fix for testIllegalFilePath

* Update SQLiteTest.cpp

* fixexes for Visual Studio and /W4 switch

* Fixed warnings for Foundation module

* Fixed warnings for JSON module

* Fixed warnings for XML module

* Fixed warnings for NET module

* Fix unfixable warning in VarHolder.h

* Fix Typo:

* Fixes for Net

* Fixes for Crypto

* Fixes for NetSSL

* Fixes for Zip

* Fixes for PDF

* Fixes for SQLite

* Fixes for SQLite

* Fixes for SQL

* Fixes for MongoDB

* Fixes for Redis

* Fixes for CppParser

* Fixes for CppUnit

* Fixes for PageCompiler

* Fixes for PostgreSQL

* Fixes for MySQL

* Fixes for SevenZip

* Last Warnings

* Fixes for ODBC

* fix std::type_info include

+++ Foundation/src/UUID.cpp
@@ -56,9 +56,9 @@ UUID::UUID(const char* uuid)
 
 
 UUID::UUID(UInt32 timeLow, UInt32 timeMid, UInt32 timeHiAndVersion, UInt16 clockSeq, UInt8 node[]):
-	_timeLow(timeLow),
-	_timeMid(timeMid),
-	_timeHiAndVersion(timeHiAndVersion),
+	_timeLow(static_cast<UInt16>(timeLow)),
+	_timeMid(static_cast<UInt16>(timeMid)),
+	_timeHiAndVersion(static_cast<UInt16>(timeHiAndVersion)),
 	_clockSeq(clockSeq)
 {
 	std::memcpy(_node, node, sizeof(_node));
@@ -183,7 +183,7 @@ bool UUID::tryParse(const std::string& uuid)
 		Int16 n2 = nibble(*it++);
 		if (n2 < 0) return false;
 
-		newUUID._node[i] = (n1 << 4) | n2;			
+		newUUID._node[i] = static_cast<UInt8>((n1 << 4) | n2);
 	}
 	swap(newUUID);
